design: search, main 페이지 컴포넌트화 및 main페이지 구현중 #16
