+++ src/app/(default)/page.tsx
@@ -1,24 +1,21 @@
 import { Metadata } from "next";
+import MainPage from "@/components/MainPage";
 
 export const metadata: Metadata = {
-  title: "pintogether",
-  description: "Generated by create next app",
-  icons: {
-    icon: [
-      {
-        url: "/logo/favicon.png",
-        media: "(prefers-color-scheme: light)",
-      },
-    ],
-  },
-};
+	title: "pintogether",
+	description: "Generated by create next app",
+	icons: {
+	  icon: [
+		{
+		  url: "/logo/favicon.png",
+		  media: "(prefers-color-scheme: light)",
+		},
+	  ],
+	},
+  };
 
 export default function Page() {
-
   return(
-    <div>
-      <h1>Home</h1>
-      가나다
-    </div>
+      <MainPage />
   );
 }
