+++ src/styles/layout/_mainPage.module.scss
@@ -1,0 +1,76 @@
+@import "../base/variables";
+
+.container {
+	display: grid;
+	grid-template-rows: 100px 100px 400px 1fr;
+	margin: 0;
+	padding: 0;
+	min-height: 100vh;
+	transition: all 0.3s ease;
+	padding-left: 7%;
+	padding-right: 7%;
+}
+
+.topper {
+	display: grid;
+	padding-top: 20px;
+	padding-bottom: 20px;
+	grid-template-columns: 200px 1fr;
+	align-items: center;
+	justify-content: flex-start;
+}
+
+.inputContainer {
+	display: grid;
+	grid-template-columns: 50px 1fr 150px;
+	width: 100%;
+	height: 50px;
+	padding-left: 20px;
+	padding-right: 20px;
+	gap: 10px;
+	border: 1px solid $color-primary;
+	border-radius: 3rem;
+	color: $color-grey-dark;
+	align-items: center;
+	justify-content: center;
+}
+
+.input {
+	font-size: $font-size-large;
+	height: 100%;
+	border-right: 1px solid $color-primary;
+}
+
+.inputButton {
+	width: 25px;
+	height: 25px;
+}
+
+.gradationBox {
+	width: 100%;
+	height: 100%;
+	background: linear-gradient(90deg, $color-secondary-light,  $color-primary-light 90%, $color-primary 140%);
+}
+
+.recommendCard{
+	padding-top: 30px;
+	padding-bottom: 30px;
+	flex:1;
+	display: grid;
+	align-items: center;
+	justify-content: center;
+	gap: 10px;
+	border-bottom: 1px solid $color-grey;
+	overflow-x: hidden;
+}
+
+.popularTop{
+	align-items: center;
+	justify-content: center;
+	place-items: center;
+}
+
+.icon {
+	width: 25px;
+	height: 25px;
+}
